refactor: service 패키지의 api 의존성 제거
